Organize scheduler checkpoint handling
